requirements.txt
# dev
black
icecream
mypy

# Google API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# NLP
marvin
trafilatura
